# tests/hospital-api-v1_2_openapi.yaml
openapi: 3.1.0
info:
  title: Hospital Management API
  version: 1.2.0
  description: |
    Patients, doctors, departments, OPD schedules, appointments, and the
    hospital information the call centre's voice agent reads to callers.

    This API is the SINGLE SOURCE for everything the voice agent says.
    Structured data (schedules, availability, prices, hours, rooms) is read
    live at the moment a caller asks. Prose (department descriptions,
    services, FAQ articles) is authored here and copied into the agent's
    search index by a sync job that polls `updated_since` (see "Sync").

    ## What changed in 1.2.0

    - NEW `Holidays` — hospital-wide closures. Availability returns a
      `holiday` status and booking refuses the date with `409 HOLIDAY`.
    - NEW `visit_type` on appointments (`new`, `follow_up`,
      `report_showing`) and an optional separate report-showing queue per
      session (`report_capacity`).
    - NEW `POST /schedules/import` — upload the OPD room-allocation
      spreadsheet (or CSV), preview the result, then apply it.
    - NEW `GET /audit-log` — who changed what, when, with before/after.
    - NEW `Hospital` profile — address, hotlines, emails, OPD and desk
      hours (replaces the voice agent's `CRM_OPENING_HOURS` setting).
    - NEW `Packages`, `Services` (including facility rooms) and `Articles`
      (FAQ and general content).
    - CHANGED `Department` — adds `description_bn` and `conditions`.
...

    ## What changed in 1.1.0

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - APIKeyHeader: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      operationId: login_auth_login_post
      security: []
      description: |
        Simple login — no registration endpoint exists.
        Users are created only via the seed script (app/seed.py).

        Returns the user's persistent API key. Send it back on every other
        request as the `X-API-Key` header (not an Authorization/Bearer token).
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { $ref: '#/components/schemas/Body_login_auth_login_post' }
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiKeyResponse' }
        '422': { $ref: '#/components/responses/ValidationError' }

  # ── Patients ─────────────────────────────────────────────────────────────
...
  /doctors/{doctor_id}/availability:
    parameters:
      - $ref: '#/components/parameters/DoctorId'
    get:
      tags: [Availability]
      summary: Doctor Availability
      operationId: get_doctor_availability
      description: |
        Weekly schedule, minus holidays, minus exceptions, minus bookings, per
        date. This is the endpoint the voice agent calls; it must never be
        cached for longer than a few seconds so a schedule change is heard on
        the next call.

        One entry per date in the range, including dates with no session.
      parameters:
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DoctorAvailability' }
              example:
                doctor_id: 12
                doctor_name: Prof. Dr. Khwaja Nazim Uddin
                days:
                  - date: '2026-09-27'
                    weekday: sunday
                    status: available
                    sessions:
                      - schedule_id: 40
                        start_time: '09:00'
                        end_time: '17:00'
                        room: '3042'
                        floor: 3rd
                        desk: A
                        capacity: 30
                        booked: 12
                        remaining: 18
                        next_serial: 13
                        report_capacity: 10
                        report_booked: 2
                        report_remaining: 8
                        now_serving: null
                  - date: '2026-09-28'
                    weekday: monday
                    status: leave
                    sessions: []
                  - date: '2026-09-30'
                    weekday: wednesday
                    status: holiday
                    holiday_name: Durga Puja
                    sessions: []
                  - date: '2026-10-02'
                    weekday: friday
                    status: no_session
                    sessions: []
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /availability:
    get:
      tags: [Availability]
      summary: Availability By Department
      operationId: get_availability
      description: |
        Every active doctor matching the filters with their availability on
        the given date range — answers "which gynae doctor sits tomorrow?".
        Doctors whose every day is `no_session` are omitted.
      parameters:
        - name: department_id
          in: query
          required: false
          schema: { type: [integer, 'null'] }
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DoctorAvailability' }
        '422': { $ref: '#/components/responses/ValidationError' }

  # ── Appointments ─────────────────────────────────────────────────────────
...
  /audit-log:
    get:
      tags: [Audit]
      summary: List Audit Entries
      operationId: list_audit_log
      description: |
        Requires `admin` scope. Every create, update, delete and import of
        doctors, departments, schedules, exceptions, holidays, packages,
        services, articles and the hospital profile, newest first. Entries
        are never edited or deleted.
      parameters:
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Limit'
        - name: entity_type
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/AuditEntityType'
              - type: 'null'
        - name: entity_id
          in: query
          required: false
          schema: { type: [integer, 'null'] }
        - name: doctor_id
          in: query
          required: false
          schema: { type: [integer, 'null'] }
          description: Entries for the doctor AND their schedules and exceptions.
        - name: user_id
          in: query
          required: false
          schema: { type: [integer, 'null'] }
        - name: from
          in: query
          required: false
          schema: { type: [string, 'null'], format: date-time }
        - name: to
          in: query
          required: false
          schema: { type: [string, 'null'], format: date-time }
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditEntry' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ── Hospital profile (NEW in 1.2.0) ──────────────────────────────────────
...
  /articles:
    get:
      tags: [Articles]
      summary: List Articles
      operationId: list_articles
      description: |
        FAQ and general content (how to book, about the hospital, careers,
        policies, directions). `agent` keys see only published articles
        unless `include_deleted=true`.
      parameters:
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Limit'
        - name: topic
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/ArticleTopic'
              - type: 'null'
        - name: service_line
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/ServiceLine'
              - type: 'null'
          description: Articles for that desk PLUS articles with no `service_line` (shared).
        - name: is_published
          in: query
          required: false
          schema: { type: [boolean, 'null'] }
          description: Admin only. `agent` keys always get published articles.
        - $ref: '#/components/parameters/UpdatedSince'
        - $ref: '#/components/parameters/IncludeDeleted'
      responses:
        '200':
          description: Successful Response
          headers:
            X-Server-Time: { $ref: '#/components/headers/XServerTime' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ArticleOut' }
    post:
      tags: [Articles]
      summary: Create Article
      operationId: create_article
      description: Requires `admin` scope.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ArticleCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArticleOut' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /articles/{article_id}:
    parameters:
      - name: article_id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Articles]
      summary: Get Article
      operationId: get_article
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArticleOut' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Articles]
      summary: Update Article
      operationId: update_article
      description: Requires `admin` scope. Partial update.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ArticleUpdate' }
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArticleOut' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Articles]
      summary: Delete Article
      operationId: delete_article
      description: Requires `admin` scope. Soft delete (sets `deleted_at`).
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Queue (NEW in 1.2.0, OPTIONAL) ───────────────────────────────────────
...
components:
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Skip:
      name: skip
      in: query
      required: false
      schema: { type: integer, default: 0, minimum: 0 }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, default: 100, minimum: 1, maximum: 500 }
    DoctorId:
      name: doctor_id
      in: path
      required: true
      schema: { type: integer }
...
    DateFrom:
      name: from
      in: query
      required: false
      schema: { type: string, format: date }
      description: Inclusive. Defaults to today (Asia/Dhaka).
    DateTo:
      name: to
      in: query
      required: false
      schema: { type: string, format: date }
      description: Inclusive. Defaults to `from + 6 days`. At most 31 days after `from`.
...
    UpdatedSince:
      name: updated_since
      in: query
      required: false
      schema: { type: [string, 'null'], format: date-time }
      description: |
        NEW in 1.2.0. Only records created, updated or deleted after this
        instant. Pass the previous response's `X-Server-Time`.
    IncludeDeleted:
      name: include_deleted
      in: query
      required: false
      schema: { type: boolean, default: false }
      description: |
        NEW in 1.2.0. Also return soft-deleted, inactive and unpublished
        records (each with `deleted_at` / `is_active` / `is_published` set),
        so a sync job can remove them from its index.
  headers:
    XServerTime:
      description: NEW in 1.2.0. Server time when the list was read; the next `updated_since`.
      schema: { type: string, format: date-time }
  responses:
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/HTTPValidationError' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: '`FORBIDDEN_SCOPE` — the API key''s scope does not allow this operation.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ErrorResponse:
      type: object
      description: NEW in 1.1.0. Body of every business-rule error.
      required: [code, detail]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        detail:
          type: string
          description: Human-readable, English. Not for parsing.
        next_available_date:
          type: [string, 'null']
          format: date
          description: |
            On `NO_SESSION`, `SESSION_FULL` and `HOLIDAY`: the doctor's next
            date (within 31 days) with a free serial, so a client can offer
            it directly.
        holiday_name:
          type: [string, 'null']
          description: NEW in 1.2.0. On `HOLIDAY`.
        candidates:
          type: [array, 'null']
          items: { type: object }
          description: Optional supporting rows (e.g. the clashing schedule on `SCHEDULE_OVERLAP`).
    ErrorCode:
      type: string
      enum:
        - NOT_FOUND
        - PATIENT_NOT_FOUND
        - DOCTOR_NOT_FOUND
        - FORBIDDEN_SCOPE
        - PAST_DATE
        - NO_SESSION
        - SESSION_REQUIRED
        - SESSION_FULL
        - DUPLICATE_BOOKING
        - DOCTOR_NOT_ACCEPTING
        - SCHEDULE_OVERLAP
        - ROOM_CONFLICT
        - HAS_DOCTORS
        - HAS_FUTURE_APPOINTMENTS
        - DUPLICATE_NAME
        - IDEMPOTENCY_KEY_REUSED
        - HOLIDAY
        - DUPLICATE_DATE
        - IMPORT_HAS_ERRORS
        - UNSUPPORTED_FILE
    Weekday:
      type: string
      enum: [saturday, sunday, monday, tuesday, wednesday, thursday, friday]
    ServiceLine:
      type: string
      description: Which call-centre desk handles this department's calls.
      enum: [appointment, emergency, support]
    ApiKeyResponse:
      type: object
      required: [api_key]
      properties:
        api_key: { type: string }
    Body_login_auth_login_post:
      type: object
      required: [username, password]
      properties:
        grant_type: { type: [string, 'null'], pattern: password }
        username: { type: string }
        password: { type: string, format: password }
        scope: { type: string, default: '' }
        client_id: { type: [string, 'null'] }
        client_secret: { type: [string, 'null'], format: password }
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items: { $ref: '#/components/schemas/ValidationError' }
    ValidationError:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            anyOf: [{ type: string }, { type: integer }]
        msg: { type: string }
        type: { type: string }

    # ── patients (unchanged shape) ─────────────────────────────────────────
...
    DepartmentSummary:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
        name_bn: { type: [string, 'null'] }

    # ── schedules (NEW) ────────────────────────────────────────────────────
    SessionPlace:
      type: object
      properties:
        room: { type: [string, 'null'], maxLength: 30, description: 'e.g. "3042"' }
        floor: { type: [string, 'null'], maxLength: 30, description: 'e.g. "3rd"' }
        desk: { type: [string, 'null'], maxLength: 10, description: 'e.g. "A"' }
...
    DayStatus:
      type: string
      enum: [available, full, leave, no_session, holiday]
      description: |
        Checked in this order:
        - `holiday` — NEW in 1.2.0. A hospital holiday with `opd_closed`.
        - `leave` — a `leave` exception, or every session `cancelled`.
        - `no_session` — no schedule row for that weekday/date.
        - `available` — at least one session with `remaining > 0`.
        - `full` — sessions exist, all with `remaining == 0`.
    AvailableSession:
      allOf:
        - $ref: '#/components/schemas/SessionPlace'
        - type: object
          required: [start_time, end_time, capacity, booked, remaining]
          properties:
            schedule_id:
              type: [integer, 'null']
              description: Null for an `extra_session` exception.
            exception_id:
              type: [integer, 'null']
              description: Set for an `extra_session` exception.
            start_time: { type: string, format: time }
            end_time: { type: string, format: time }
            capacity: { type: integer }
            booked: { type: integer, description: 'Appointments not `cancelled` (a `completed` one still used its serial).' }
            remaining: { type: integer }
            next_serial:
              type: [integer, 'null']
              description: The serial the next booking would get. Null when full.
            report_capacity: { type: integer, description: NEW in 1.2.0. 0 when report visits use regular serials. }
            report_booked: { type: integer, description: NEW in 1.2.0. }
            report_remaining: { type: integer, description: NEW in 1.2.0. }
            now_serving:
              type: [integer, 'null']
              description: NEW in 1.2.0, OPTIONAL queue feature. Today only; null otherwise or when not tracked.
    AvailabilityDay:
      type: object
      required: [date, weekday, status, sessions]
      properties:
        date: { type: string, format: date }
        weekday: { $ref: '#/components/schemas/Weekday' }
        status: { $ref: '#/components/schemas/DayStatus' }
        holiday_name:
          type: [string, 'null']
          description: NEW in 1.2.0. Set when `status` is `holiday`.
        sessions:
          type: array
          items: { $ref: '#/components/schemas/AvailableSession' }
          description: Empty unless status is `available` or `full`.
    DoctorAvailability:
      type: object
      required: [doctor_id, doctor_name, days]
      properties:
        doctor_id: { type: integer }
        doctor_name: { type: string }
        doctor_name_bn: { type: [string, 'null'] }
        department:
          anyOf:
            - $ref: '#/components/schemas/DepartmentSummary'
            - type: 'null'
        accepting_appointments: { type: boolean }
        days:
          type: array
          items: { $ref: '#/components/schemas/AvailabilityDay' }

    # ── appointments ───────────────────────────────────────────────────────
...
    AuditEntityType:
      type: string
      enum: [doctor, department, schedule, exception, holiday, package, service, article, hospital]
    AuditEntry:
      type: object
      required: [id, at, user_id, action, entity_type, entity_id]
      properties:
        id: { type: integer }
        at: { type: string, format: date-time }
        user_id: { type: integer }
        username: { type: string }
        action: { type: string, enum: [create, update, delete, import] }
        entity_type: { $ref: '#/components/schemas/AuditEntityType' }
        entity_id: { type: integer }
        doctor_id:
          type: [integer, 'null']
          description: For doctor, schedule and exception entries.
        before:
          type: [object, 'null']
          description: The record before the change; null on create.
        after:
          type: [object, 'null']
          description: The record after the change; null on delete.

    # ── hospital (NEW in 1.2.0) ────────────────────────────────────────────
...
    ArticleTopic:
      type: string
      enum: [booking, about, directions, policy, career, other]
    ArticleCreate:
      type: object
      required: [title, topic]
      properties:
        title: { type: string, maxLength: 200, description: 'e.g. "How to book an appointment"' }
        title_bn: { type: [string, 'null'], maxLength: 200 }
        topic: { $ref: '#/components/schemas/ArticleTopic' }
        service_line:
          anyOf:
            - $ref: '#/components/schemas/ServiceLine'
            - type: 'null'
          description: The desk this answers for. Null means every desk (shared).
        questions:
          type: array
          items: { type: string, maxLength: 200 }
          default: []
          description: |
            How callers actually ask it, several phrasings, either language:
            ["সিরিয়াল কীভাবে নেব?", "অ্যাপয়েন্টমেন্ট কীভাবে করা যায়?"].
            Improves search matching.
        body:
          type: [string, 'null']
          description: The answer in English, short and speakable.
        body_bn:
          type: [string, 'null']
          description: The answer in Bangla. At least one of `body`, `body_bn` is required.
        is_published: { type: boolean, default: false }
    ArticleUpdate:
      type: object
      description: Every field optional; omitted fields are unchanged.
      properties:
        title: { type: [string, 'null'], maxLength: 200 }
        title_bn: { type: [string, 'null'], maxLength: 200 }
        topic:
          anyOf:
            - $ref: '#/components/schemas/ArticleTopic'
            - type: 'null'
        service_line:
          anyOf:
            - $ref: '#/components/schemas/ServiceLine'
            - type: 'null'
        questions: { type: [array, 'null'], items: { type: string, maxLength: 200 } }
        body: { type: [string, 'null'] }
        body_bn: { type: [string, 'null'] }
        is_published: { type: [boolean, 'null'] }
    ArticleOut:
      allOf:
        - $ref: '#/components/schemas/ArticleCreate'
        - type: object
          required: [id, created_at]
          properties:
            id: { type: integer }
            created_at: { type: string, format: date-time }
            updated_at: { type: [string, 'null'], format: date-time }
            deleted_at: { type: [string, 'null'], format: date-time }

    # ── queue (NEW in 1.2.0, OPTIONAL) ─────────────────────────────────────
...